+++ vllm/v1/attention/backends/mla/rocm_aiter_mla_sparse.py
@@ -2,7 +2,7 @@
 # SPDX-FileCopyrightText: Copyright contributors to the vLLM project
 
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, ClassVar, cast
+from typing import TYPE_CHECKING, ClassVar
 
 import numpy as np
 import torch
@@ -11,7 +11,6 @@ from vllm import _custom_ops as ops
 from vllm._aiter_ops import rocm_aiter_ops
 from vllm.config import VllmConfig
 from vllm.config.cache import CacheDType
-from vllm.forward_context import get_forward_context
 from vllm.logger import init_logger
 from vllm.model_executor.layers.attention.mla_attention import (
     get_mla_dims,
@@ -28,33 +27,12 @@ from vllm.v1.attention.backend import (
     MultipleOf,
     SparseMLAAttentionImpl,
 )
-from vllm.v1.attention.backends.mla.flashmla_sparse import (
-    DeepseekV4FlashMLASparseBackend,
-    FlashMLASparseMetadata,
-    FlashMLASparseMetadataBuilder,
-)
 from vllm.v1.attention.backends.mla.rocm_aiter_mla import (
     AiterMLAHelper,
 )
-from vllm.v1.attention.backends.mla.sparse_swa import (
-    DeepseekSparseSWAMetadata,
-    DeepseekSparseSWAMetadataBuilder,
-)
-from vllm.v1.attention.ops.deepseek_v4_ops import (
-    dequantize_and_gather_k_cache,
-)
-from vllm.v1.attention.ops.rocm_aiter_mla_sparse import (
-    build_ragged_indices_from_dense,
-    rocm_sparse_attn_decode,
-    rocm_sparse_attn_prefill,
-)
 from vllm.v1.kv_cache_interface import AttentionSpec
-from vllm.v1.worker.workspace import current_workspace_manager
 
 if TYPE_CHECKING:
-    from vllm.model_executor.layers.deepseek_v4_attention import (
-        DeepseekV4MLAAttention,
-    )
     from vllm.model_executor.models.deepseek_v2 import Indexer
 logger = init_logger(__name__)
 
@@ -724,650 +702,3 @@ class ROCMAiterMLASparseImpl(SparseMLAAttentionImpl[ROCMAiterMLASparseMetadata])
         )
 
         return attn_out, None
-
-
-def _build_indptr_from_lengths(lengths: torch.Tensor) -> torch.Tensor:
-    lengths = lengths.to(dtype=torch.int32).contiguous()
-    indptr = torch.empty(lengths.shape[0] + 1, dtype=torch.int32, device=lengths.device)
-    indptr[0] = 0
-    torch.cumsum(lengths, dim=0, out=indptr[1:])
-    return indptr
-
-
-@triton.jit
-def _compute_topk_lens_kernel(
-    topk_lens_ptr,
-    topk_indices_ptr,
-    topk_indices_stride,
-    topk,
-    is_valid_token_ptr,
-    TRITON_BLOCK_SIZE: tl.constexpr,
-):
-    token_idx = tl.program_id(0)
-    is_valid_token = tl.load(is_valid_token_ptr + token_idx)
-
-    count = tl.zeros((), dtype=tl.int32)
-    for i in range(0, topk, TRITON_BLOCK_SIZE):
-        offset = i + tl.arange(0, TRITON_BLOCK_SIZE)
-        mask = offset < topk
-        local_idx = tl.load(
-            topk_indices_ptr + token_idx * topk_indices_stride + offset,
-            mask=mask,
-            other=-1,
-        )
-        count += tl.sum((local_idx >= 0).to(tl.int32), axis=0)
-
-    tl.store(topk_lens_ptr + token_idx, tl.where(is_valid_token, count, 0))
-
-
-@triton.jit
-def _pack_global_topk_ragged_kernel(
-    global_topk_ragged_ptr,
-    topk_indptr_ptr,
-    topk_indices_ptr,
-    topk_indices_stride,
-    token_to_req_indices_ptr,
-    block_table_ptr,
-    block_table_stride,
-    block_size,
-    topk,
-    BLOCK_SIZE: tl.constexpr,
-):
-    token_idx = tl.program_id(0)
-    block_idx = tl.program_id(1)
-    offset = block_idx * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
-
-    out_start = tl.load(topk_indptr_ptr + token_idx)
-    out_end = tl.load(topk_indptr_ptr + token_idx + 1)
-    out_len = out_end - out_start
-    if block_idx * BLOCK_SIZE >= out_len:
-        return
-
-    req_idx = tl.load(token_to_req_indices_ptr + token_idx)
-    mask = (offset < out_len) & (offset < topk)
-    local_idx = tl.load(
-        topk_indices_ptr + token_idx * topk_indices_stride + offset,
-        mask=mask,
-        other=-1,
-    )
-    valid = mask & (local_idx >= 0)
-    block_indices = local_idx // block_size
-    block_numbers = tl.load(
-        block_table_ptr + req_idx * block_table_stride + block_indices,
-        mask=valid,
-        other=0,
-    )
-    block_offsets = local_idx % block_size
-    slot_ids = tl.where(valid, block_numbers * block_size + block_offsets, -1)
-    tl.store(global_topk_ragged_ptr + out_start + offset, slot_ids, mask=mask)
-
-
-def compute_global_topk_ragged_indices_and_indptr(
-    topk_indices: torch.Tensor,
-    token_to_req_indices: torch.Tensor,
-    block_table: torch.Tensor,
-    block_size: int,
-    is_valid_token: torch.Tensor,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    topk_indices = topk_indices.reshape(topk_indices.shape[0], -1).contiguous()
-    num_tokens = topk_indices.shape[0]
-    topk = topk_indices.shape[1]
-
-    topk_lens = torch.empty(num_tokens, dtype=torch.int32, device=topk_indices.device)
-    _compute_topk_lens_kernel[(num_tokens,)](
-        topk_lens,
-        topk_indices,
-        topk_indices.stride(0),
-        topk,
-        is_valid_token,
-        TRITON_BLOCK_SIZE=1024,
-    )
-
-    topk_indptr = _build_indptr_from_lengths(topk_lens)
-    global_topk_ragged = torch.empty(
-        int(topk_indptr[-1].item()),
-        dtype=torch.int32,
-        device=topk_indices.device,
-    )
-    if global_topk_ragged.numel() > 0:
-        block = 128
-        _pack_global_topk_ragged_kernel[(num_tokens, triton.cdiv(topk, block))](
-            global_topk_ragged,
-            topk_indptr,
-            topk_indices,
-            topk_indices.stride(0),
-            token_to_req_indices,
-            block_table,
-            block_table.stride(0),
-            block_size,
-            topk,
-            BLOCK_SIZE=block,
-        )
-    return global_topk_ragged, topk_indptr, topk_lens
-
-
-@triton.jit
-def _compute_combined_lens_kernel(
-    combined_lens_ptr,
-    query_start_loc_ptr,
-    seq_lens_ptr,
-    TOP_K: tl.constexpr,
-    COMPRESS_RATIO: tl.constexpr,
-    WINDOW_SIZE: tl.constexpr,
-):
-    batch_idx = tl.program_id(0)
-    worker_id = tl.program_id(1)
-    num_workers = tl.num_programs(1)
-
-    base = tl.load(query_start_loc_ptr)
-    query_start = tl.load(query_start_loc_ptr + batch_idx) - base
-    query_end = tl.load(query_start_loc_ptr + batch_idx + 1) - base
-    query_len = query_end - query_start
-    seq_len = tl.load(seq_lens_ptr + batch_idx)
-    start_pos = seq_len - query_len
-
-    for token_idx in range(query_start + worker_id, query_end, num_workers):
-        token_idx_in_query = token_idx - query_start
-        pos = start_pos + token_idx_in_query
-        topk_len = tl.minimum((pos + 1) // COMPRESS_RATIO, TOP_K)
-        swa_len = tl.minimum(pos + 1, WINDOW_SIZE)
-        tl.store(combined_lens_ptr + token_idx, topk_len + swa_len)
-
-
-@triton.jit
-def _combine_topk_swa_indices_ragged_kernel(
-    combined_ragged_ptr,
-    combined_indptr_ptr,
-    topk_indices_ptr,
-    topk_indices_stride,
-    query_start_loc_ptr,
-    seq_lens_ptr,
-    gather_lens_ptr,
-    M,
-    N,
-    topk_width,
-    TOP_K: tl.constexpr,
-    COMPRESS_RATIO: tl.constexpr,
-    WINDOW_SIZE: tl.constexpr,
-    BLOCK_SIZE: tl.constexpr,
-):
-    batch_idx = tl.program_id(0)
-    worker_id = tl.program_id(1)
-    block_idx = tl.program_id(2)
-    num_workers = tl.num_programs(1)
-
-    base = tl.load(query_start_loc_ptr)
-    query_start = tl.load(query_start_loc_ptr + batch_idx) - base
-    query_end = tl.load(query_start_loc_ptr + batch_idx + 1) - base
-    query_len = query_end - query_start
-    seq_len = tl.load(seq_lens_ptr + batch_idx)
-    gather_len = tl.load(gather_lens_ptr + batch_idx)
-    start_pos = seq_len - query_len
-    gather_start = seq_len - gather_len
-
-    for token_idx in range(query_start + worker_id, query_end, num_workers):
-        token_idx_in_query = token_idx - query_start
-        pos = start_pos + token_idx_in_query
-        topk_len = tl.minimum((pos + 1) // COMPRESS_RATIO, TOP_K)
-        swa_len = tl.minimum(pos + 1, WINDOW_SIZE)
-        combined_len = topk_len + swa_len
-
-        offset = block_idx * BLOCK_SIZE + tl.arange(0, BLOCK_SIZE)
-        if block_idx * BLOCK_SIZE < combined_len:
-            out_start = tl.load(combined_indptr_ptr + token_idx)
-            topk_mask = (offset < topk_len) & (offset < topk_width)
-            topk_vals = tl.load(
-                topk_indices_ptr + token_idx * topk_indices_stride + offset,
-                mask=topk_mask,
-                other=-1,
-            )
-            tl.store(
-                combined_ragged_ptr + out_start + offset,
-                topk_vals + M * batch_idx,
-                mask=topk_mask,
-            )
-
-            swa_offset = offset - topk_len
-            swa_mask = (offset >= topk_len) & (swa_offset < swa_len)
-            tl.store(
-                combined_ragged_ptr + out_start + offset,
-                M * batch_idx + N + swa_offset + pos - swa_len + 1 - gather_start,
-                mask=swa_mask,
-            )
-
-
-def combine_topk_swa_indices_ragged(
-    topk_indices: torch.Tensor,
-    query_start_loc: torch.Tensor,
-    seq_lens: torch.Tensor,
-    gather_lens: torch.Tensor,
-    window_size: int,
-    compress_ratio: int,
-    topk: int,
-    M: int,
-    N: int,
-) -> tuple[torch.Tensor, torch.Tensor, torch.Tensor]:
-    topk_indices = topk_indices.reshape(topk_indices.shape[0], -1).contiguous()
-    num_tokens = topk_indices.shape[0]
-    num_reqs = seq_lens.shape[0]
-    combined_lens = torch.empty(
-        num_tokens, dtype=torch.int32, device=topk_indices.device
-    )
-
-    num_workers = 128
-    _compute_combined_lens_kernel[(num_reqs, num_workers)](
-        combined_lens,
-        query_start_loc,
-        seq_lens,
-        TOP_K=topk,
-        COMPRESS_RATIO=compress_ratio,
-        WINDOW_SIZE=window_size,
-    )
-
-    combined_indptr = _build_indptr_from_lengths(combined_lens)
-    combined_ragged = torch.empty(
-        int(combined_indptr[-1].item()),
-        dtype=torch.int32,
-        device=topk_indices.device,
-    )
-    if combined_ragged.numel() > 0:
-        block = 128
-        _combine_topk_swa_indices_ragged_kernel[
-            (num_reqs, num_workers, triton.cdiv(topk + window_size, block))
-        ](
-            combined_ragged,
-            combined_indptr,
-            topk_indices,
-            topk_indices.stride(0),
-            query_start_loc,
-            seq_lens,
-            gather_lens,
-            M,
-            N,
-            topk_indices.shape[-1],
-            TOP_K=topk,
-            COMPRESS_RATIO=compress_ratio,
-            WINDOW_SIZE=window_size,
-            BLOCK_SIZE=block,
-        )
-    return combined_ragged, combined_indptr, combined_lens
-
-
-@dataclass
-class DeepseekV4ROCMAiterMLASparseMetadata(FlashMLASparseMetadata):
-    """ROCm-specific DeepSeek V4 metadata carrying ragged decode topk."""
-
-    c128a_decode_topk_ragged_indices: torch.Tensor | None = None
-    c128a_decode_topk_ragged_indptr: torch.Tensor | None = None
-
-
-@dataclass
-class DeepseekV4ROCMAiterSparseSWAMetadata(DeepseekSparseSWAMetadata):
-    decode_swa_ragged_indices: torch.Tensor | None = None
-    decode_swa_ragged_indptr: torch.Tensor | None = None
-
-
-class DeepseekV4ROCMAiterMLASparseMetadataBuilder(FlashMLASparseMetadataBuilder):
-    def build(
-        self,
-        common_prefix_len: int,
-        common_attn_metadata: CommonAttentionMetadata,
-        fast_build: bool = False,
-    ) -> DeepseekV4ROCMAiterMLASparseMetadata:
-        base = super().build(
-            common_prefix_len=common_prefix_len,
-            common_attn_metadata=common_attn_metadata,
-            fast_build=fast_build,
-        )
-
-        ragged_indices = None
-        ragged_indptr = None
-        dense_decode = base.c128a_global_decode_topk_indices
-        decode_lens = base.c128a_decode_topk_lens
-        if dense_decode is not None and decode_lens is not None:
-            ragged_indices, ragged_indptr = build_ragged_indices_from_dense(
-                dense_decode.reshape(dense_decode.shape[0], -1),
-                decode_lens,
-            )
-
-        return DeepseekV4ROCMAiterMLASparseMetadata(
-            **vars(base),
-            c128a_decode_topk_ragged_indices=ragged_indices,
-            c128a_decode_topk_ragged_indptr=ragged_indptr,
-        )
-
-
-class DeepseekV4ROCMAiterSparseSWAMetadataBuilder(DeepseekSparseSWAMetadataBuilder):
-    def build(
-        self,
-        common_prefix_len: int,
-        common_attn_metadata: CommonAttentionMetadata,
-        fast_build: bool = False,
-    ) -> DeepseekV4ROCMAiterSparseSWAMetadata:
-        base = super().build(
-            common_prefix_len=common_prefix_len,
-            common_attn_metadata=common_attn_metadata,
-            fast_build=fast_build,
-        )
-
-        ragged_indices = None
-        ragged_indptr = None
-        if (
-            base.num_decode_tokens > 0
-            and base.decode_swa_indices is not None
-            and base.decode_swa_lens is not None
-        ):
-            ragged_indices, ragged_indptr = build_ragged_indices_from_dense(
-                base.decode_swa_indices.reshape(base.num_decode_tokens, -1),
-                base.decode_swa_lens,
-            )
-
-        return DeepseekV4ROCMAiterSparseSWAMetadata(
-            **vars(base),
-            decode_swa_ragged_indices=ragged_indices,
-            decode_swa_ragged_indptr=ragged_indptr,
-        )
-
-
-class DeepseekV4ROCMAiterMLASparseImpl(
-    SparseMLAAttentionImpl[DeepseekV4ROCMAiterMLASparseMetadata]
-):
-    """ROCm sparse MLA implementation used by DeepSeek V4's custom MLA layer."""
-
-    _PREFILL_CHUNK_SIZE = 4
-
-    def __init__(
-        self,
-        num_heads: int,
-        head_size: int,
-        scale: float,
-        num_kv_heads: int,
-        alibi_slopes: list[float] | None,
-        sliding_window: int | None,
-        kv_cache_dtype: str,
-        logits_soft_cap: float | None,
-        attn_type: str,
-        kv_sharing_target_layer_name: str | None,
-        **_: object,
-    ) -> None:
-        self.num_heads = num_heads
-        self.head_size = head_size
-        self.scale = float(scale)
-        self.num_kv_heads = num_kv_heads
-        self.kv_cache_dtype = kv_cache_dtype
-
-    def forward_mqa(
-        self,
-        q: torch.Tensor | tuple[torch.Tensor, torch.Tensor],
-        kv_c_and_k_pe_cache: torch.Tensor,
-        attn_metadata: DeepseekV4ROCMAiterMLASparseMetadata,
-        layer: AttentionLayer,
-    ) -> tuple[torch.Tensor, torch.Tensor | None]:
-        raise NotImplementedError(
-            "DeepseekV4ROCMAiterMLASparseImpl is driven by "
-            "DeepseekV4MLAAttention.forward."
-        )
-
-    @classmethod
-    def forward(
-        cls,
-        layer: "DeepseekV4MLAAttention",
-        q: torch.Tensor,
-        kv: torch.Tensor,
-        positions: torch.Tensor,
-        output: torch.Tensor,
-    ) -> None:
-        assert output.shape == q.shape, (
-            f"output buffer shape {output.shape} must match q shape {q.shape}"
-        )
-        assert output.dtype == q.dtype, (
-            f"output buffer dtype {output.dtype} must match q dtype {q.dtype}"
-        )
-
-        forward_context = get_forward_context()
-        attn_metadata = forward_context.attn_metadata
-        assert isinstance(attn_metadata, dict)
-        rocm_metadata = cast(
-            DeepseekV4ROCMAiterMLASparseMetadata | None,
-            attn_metadata.get(layer.prefix),
-        )
-        swa_metadata = cast(
-            DeepseekV4ROCMAiterSparseSWAMetadata | None,
-            attn_metadata.get(layer.swa_cache_layer.prefix),
-        )
-        assert swa_metadata is not None
-
-        swa_only = layer.compress_ratio <= 1
-        self_kv_cache = layer.kv_cache if not swa_only else None
-        swa_kv_cache = layer.swa_cache_layer.kv_cache
-
-        num_decodes = swa_metadata.num_decodes
-        num_prefills = swa_metadata.num_prefills
-        num_decode_tokens = swa_metadata.num_decode_tokens
-
-        if num_prefills > 0:
-            cls._forward_prefill(
-                layer=layer,
-                q=q[num_decode_tokens:],
-                positions=positions[num_decode_tokens:],
-                compressed_k_cache=self_kv_cache,
-                swa_k_cache=swa_kv_cache,
-                output=output[num_decode_tokens:],
-                attn_metadata=rocm_metadata,
-                swa_metadata=swa_metadata,
-            )
-        if num_decodes > 0:
-            cls._forward_decode(
-                layer=layer,
-                q=q[:num_decode_tokens],
-                kv_cache=self_kv_cache,
-                swa_metadata=swa_metadata,
-                attn_metadata=rocm_metadata,
-                swa_only=swa_only,
-                output=output[:num_decode_tokens],
-            )
-
-    @classmethod
-    def _forward_decode(
-        cls,
-        layer: "DeepseekV4MLAAttention",
-        q: torch.Tensor,
-        kv_cache: torch.Tensor | None,
-        swa_metadata: DeepseekV4ROCMAiterSparseSWAMetadata,
-        attn_metadata: DeepseekV4ROCMAiterMLASparseMetadata | None,
-        swa_only: bool,
-        output: torch.Tensor,
-    ) -> None:
-        num_decodes = swa_metadata.num_decodes
-        num_decode_tokens = swa_metadata.num_decode_tokens
-
-        topk_indices = None
-        topk_lens = None
-        topk_ragged_indices = None
-        topk_ragged_indptr = None
-        if not swa_only:
-            assert attn_metadata is not None
-            assert swa_metadata.is_valid_token is not None
-            block_size = attn_metadata.block_size // layer.compress_ratio
-            is_valid = swa_metadata.is_valid_token[:num_decode_tokens]
-            if layer.compress_ratio == 4:
-                assert layer.topk_indices_buffer is not None
-                (
-                    topk_ragged_indices,
-                    topk_ragged_indptr,
-                    topk_lens,
-                ) = compute_global_topk_ragged_indices_and_indptr(
-                    layer.topk_indices_buffer[:num_decode_tokens],
-                    swa_metadata.token_to_req_indices,
-                    attn_metadata.block_table[:num_decodes],
-                    block_size,
-                    is_valid,
-                )
-            else:
-                topk_indices = attn_metadata.c128a_global_decode_topk_indices
-                topk_lens = attn_metadata.c128a_decode_topk_lens
-                topk_ragged_indices = attn_metadata.c128a_decode_topk_ragged_indices
-                topk_ragged_indptr = attn_metadata.c128a_decode_topk_ragged_indptr
-
-        rocm_sparse_attn_decode(
-            q=q,
-            kv_cache=kv_cache,
-            swa_k_cache=layer.swa_cache_layer.kv_cache,
-            swa_only=swa_only,
-            topk_indices=topk_indices,
-            topk_lens=topk_lens,
-            swa_indices=swa_metadata.decode_swa_indices,
-            swa_lens=swa_metadata.decode_swa_lens,
-            swa_ragged_indices=swa_metadata.decode_swa_ragged_indices,
-            swa_ragged_indptr=swa_metadata.decode_swa_ragged_indptr,
-            topk_ragged_indices=topk_ragged_indices,
-            topk_ragged_indptr=topk_ragged_indptr,
-            attn_sink=layer.attn_sink,
-            scale=layer.scale,
-            head_dim=layer.head_dim,
-            nope_head_dim=layer.nope_head_dim,
-            rope_head_dim=layer.rope_head_dim,
-            output=output,
-        )
-
-    @classmethod
-    def _forward_prefill(
-        cls,
-        layer: "DeepseekV4MLAAttention",
-        q: torch.Tensor,
-        positions: torch.Tensor,
-        compressed_k_cache: torch.Tensor | None,
-        swa_k_cache: torch.Tensor,
-        output: torch.Tensor,
-        attn_metadata: DeepseekV4ROCMAiterMLASparseMetadata | None,
-        swa_metadata: DeepseekV4ROCMAiterSparseSWAMetadata,
-    ) -> None:
-        swa_only = attn_metadata is None
-
-        num_prefills = swa_metadata.num_prefills
-        num_prefill_tokens = swa_metadata.num_prefill_tokens
-        num_decodes = swa_metadata.num_decodes
-        num_decode_tokens = swa_metadata.num_decode_tokens
-
-        seq_lens = swa_metadata.prefill_seq_lens
-        gather_lens = swa_metadata.prefill_gather_lens
-        assert seq_lens is not None
-        assert gather_lens is not None
-
-        query_start_loc_cpu = swa_metadata.query_start_loc_cpu
-        query_start_loc = swa_metadata.query_start_loc
-        assert query_start_loc_cpu is not None
-        assert query_start_loc is not None
-        prefill_token_base = query_start_loc_cpu[num_decodes]
-
-        if not swa_only:
-            if layer.compress_ratio == 4:
-                assert layer.topk_indices_buffer is not None
-                topk_indices = layer.topk_indices_buffer[num_decode_tokens:]
-                topk_indices = topk_indices[:num_prefill_tokens]
-            else:
-                assert attn_metadata is not None
-                topk_indices = attn_metadata.c128a_prefill_topk_indices
-            assert topk_indices is not None
-            top_k = topk_indices.shape[-1]
-            N = (layer.max_model_len + layer.compress_ratio - 1) // layer.compress_ratio
-        else:
-            assert layer.topk_indices_buffer is not None
-            topk_indices = layer.topk_indices_buffer[num_decode_tokens:]
-            top_k = 0
-            N = 0
-
-        M = N + layer.window_size + layer.max_num_batched_tokens
-        num_chunks = (num_prefills + cls._PREFILL_CHUNK_SIZE - 1) // (
-            cls._PREFILL_CHUNK_SIZE
-        )
-
-        workspace_manager = current_workspace_manager()
-        kv = workspace_manager.get_simultaneous(
-            ((cls._PREFILL_CHUNK_SIZE, M, q.shape[-1]), torch.bfloat16),
-        )[0]
-        for chunk_idx in range(num_chunks):
-            chunk_start = chunk_idx * cls._PREFILL_CHUNK_SIZE
-            chunk_end = min(chunk_start + cls._PREFILL_CHUNK_SIZE, num_prefills)
-            chunk_size = chunk_end - chunk_start
-            if not swa_only:
-                assert attn_metadata is not None
-                assert compressed_k_cache is not None
-                block_table = attn_metadata.block_table[num_decodes:]
-                dequantize_and_gather_k_cache(
-                    kv[:chunk_size],
-                    compressed_k_cache,
-                    seq_lens=seq_lens[chunk_start:chunk_end] // layer.compress_ratio,
-                    gather_lens=None,
-                    block_table=block_table[chunk_start:chunk_end],
-                    block_size=attn_metadata.block_size // layer.compress_ratio,
-                    offset=0,
-                )
-
-            swa_block_table = swa_metadata.block_table[num_decodes:]
-            dequantize_and_gather_k_cache(
-                kv[:chunk_size],
-                swa_k_cache,
-                seq_lens=seq_lens[chunk_start:chunk_end],
-                gather_lens=gather_lens[chunk_start:chunk_end],
-                block_table=swa_block_table[chunk_start:chunk_end],
-                block_size=swa_metadata.block_size,
-                offset=N,
-            )
-
-            query_start = (
-                query_start_loc_cpu[num_decodes + chunk_start] - prefill_token_base
-            )
-            query_end = (
-                query_start_loc_cpu[num_decodes + chunk_end] - prefill_token_base
-            )
-
-            combined_ragged_indices, combined_ragged_indptr, combined_lens = (
-                combine_topk_swa_indices_ragged(
-                    topk_indices[query_start:query_end],
-                    query_start_loc[
-                        num_decodes + chunk_start : num_decodes + chunk_end + 1
-                    ],
-                    seq_lens[chunk_start:chunk_end],
-                    gather_lens[chunk_start:chunk_end],
-                    layer.window_size,
-                    layer.compress_ratio,
-                    top_k,
-                    M,
-                    N,
-                )
-            )
-            rocm_sparse_attn_prefill(
-                q=q[query_start:query_end],
-                kv=kv.view(-1, 1, q.shape[-1]),
-                indices=torch.empty(
-                    q[query_start:query_end].shape[0],
-                    1,
-                    0,
-                    dtype=torch.int32,
-                    device=q.device,
-                ),
-                topk_length=combined_lens,
-                scale=layer.scale,
-                head_dim=layer.head_dim,
-                nope_head_dim=layer.nope_head_dim,
-                rope_head_dim=layer.rope_head_dim,
-                attn_sink=layer.attn_sink,
-                output=output[query_start:query_end],
-                ragged_indices=combined_ragged_indices,
-                ragged_indptr=combined_ragged_indptr,
-            )
-
-
-class DeepseekV4ROCMAiterMLASparseBackend(DeepseekV4FlashMLASparseBackend):
-    @staticmethod
-    def get_name() -> str:
-        return "ROCM_V4_FLASHMLA_SPARSE"
-
-    @staticmethod
-    def get_builder_cls() -> type["DeepseekV4ROCMAiterMLASparseMetadataBuilder"]:
-        return DeepseekV4ROCMAiterMLASparseMetadataBuilder
-
-    @staticmethod
-    def get_impl_cls() -> type["DeepseekV4ROCMAiterMLASparseImpl"]:
-        return DeepseekV4ROCMAiterMLASparseImpl
